+++ Tests/IGListAdapterTests.m
@@ -1089,4 +1089,129 @@
     XCTAssertEqual(size.height, 96);
 }
 
+- (void)test_whenInsertingAtBeginning_thatAllSectionControllerIndexesUpdateCorrectly_forInsertAtHead {
+    NSNumber *zero = @0;
+    NSNumber *one = @1;
+    NSNumber *two = @2;
+    NSNumber *three = @3;
+    self.dataSource.objects = @[one, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    IGListSectionController *controller1a = [self.adapter sectionControllerForObject:one];
+    XCTAssertEqual(controller1a.sectionIndex, 0);
+    XCTAssertTrue(controller1a.isFirstSection);
+
+    XCTAssertEqual([self.adapter sectionControllerForObject:two].sectionIndex, 1);
+    XCTAssertEqual([self.adapter sectionControllerForObject:three].sectionIndex, 2);
+    XCTAssertTrue([self.adapter sectionControllerForObject:three].isLastSection);
+
+    self.dataSource.objects = @[zero, one, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    IGListSectionController *controller0 = [self.adapter sectionControllerForObject:zero];
+    XCTAssertEqual(controller0.sectionIndex, 0);
+    XCTAssertTrue(controller0.isFirstSection);
+
+    IGListSectionController *controller1b = [self.adapter sectionControllerForObject:one];
+    XCTAssertEqual(controller1b.sectionIndex, 1);
+    XCTAssertFalse(controller1b.isFirstSection);
+
+    XCTAssertEqual([self.adapter sectionControllerForObject:two].sectionIndex, 2);
+    XCTAssertEqual([self.adapter sectionControllerForObject:three].sectionIndex, 3);
+    XCTAssertTrue([self.adapter sectionControllerForObject:three].isLastSection);
+}
+
+- (void)test_whenRemovingFromHead_thatAllSectionControllerIndexesUpdateCorrectly_RemovedSectionControllerIsNotFound {
+    NSNumber *zero = @0;
+    NSNumber *one = @1;
+    NSNumber *two = @2;
+    NSNumber *three = @3;
+    self.dataSource.objects = @[zero, one, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    IGListSectionController *zeroController = [self.adapter sectionControllerForSection:0];
+    XCTAssertEqual(zeroController.sectionIndex, 0);
+    XCTAssertTrue(zeroController.isFirstSection);
+
+    IGListSectionController *oneController = [self.adapter sectionControllerForSection:1];
+    XCTAssertEqual(oneController.sectionIndex, 1);
+    XCTAssertFalse(oneController.isFirstSection);
+
+    IGListSectionController *threeController = [self.adapter sectionControllerForSection:3];
+    XCTAssertEqual(threeController.sectionIndex, 3);
+    XCTAssertTrue(threeController.isLastSection);
+
+    self.dataSource.objects = @[one, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    XCTAssertEqual(zeroController.sectionIndex, NSNotFound);
+    XCTAssertFalse(zeroController.isFirstSection);
+
+    XCTAssertEqual(oneController.sectionIndex, 0);
+    XCTAssertTrue(oneController.isFirstSection);
+
+    XCTAssertEqual(threeController.sectionIndex, 2);
+    XCTAssertTrue(threeController.isLastSection);
+}
+
+- (void)test_whenRemovingFromMiddle_thatAllSectionControllerIndexesUpdateCorrectly_removedSectionControllerIsNotFound {
+    NSNumber *zero = @0;
+    NSNumber *one = @1;
+    NSNumber *two = @2;
+    NSNumber *three = @3;
+    self.dataSource.objects = @[zero, one, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    IGListSectionController *zeroController = [self.adapter sectionControllerForSection:0];
+    XCTAssertEqual(zeroController.sectionIndex, 0);
+    XCTAssertTrue(zeroController.isFirstSection);
+
+    IGListSectionController *oneController = [self.adapter sectionControllerForSection:1];
+    XCTAssertEqual(oneController.sectionIndex, 1);
+    XCTAssertFalse(oneController.isFirstSection);
+
+    IGListSectionController *threeController = [self.adapter sectionControllerForSection:3];
+    XCTAssertEqual(threeController.sectionIndex, 3);
+    XCTAssertTrue(threeController.isLastSection);
+
+    self.dataSource.objects = @[zero, two, three];
+    [self.adapter performUpdatesAnimated:NO completion:nil];
+
+    XCTAssertEqual(zeroController.sectionIndex, 0);
+    XCTAssertTrue(zeroController.isFirstSection);
+
+    XCTAssertEqual(oneController.sectionIndex, NSNotFound);
+    XCTAssertFalse(oneController.isFirstSection);
+
+    XCTAssertEqual(threeController.sectionIndex, 2);
+    XCTAssertTrue(threeController.isLastSection);
+}
+
+- (void)test_withStrongRefToSectionController_thatAdapterSectionIndexIsZero_thatSectionControllerIndexDoesNotChange {
+    IGListSectionController *sc = nil;
+
+    // hold a weak reference to simulate what would happen to the collectionContext object on a section controller
+    // if the section controller were held strongly by an async block and the rest of the infra was deallocated
+    __weak IGListAdapter *wAdapter = nil;
+
+    @autoreleasepool {
+        IGListTestAdapterDataSource *dataSource = [IGListTestAdapterDataSource new];
+        IGListReloadDataUpdater *updater = [IGListReloadDataUpdater new];
+        IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater
+                                                         viewController:nil];
+        adapter.dataSource = dataSource;
+        adapter.collectionView = self.collectionView;
+        wAdapter = adapter;
+
+        dataSource.objects = @[@0, @1, @2];
+        [adapter performUpdatesAnimated:NO completion:nil];
+
+        sc = [adapter sectionControllerForSection:1];
+        XCTAssertEqual(sc.sectionIndex, 1);
+    }
+
+    XCTAssertEqual(sc.sectionIndex, 1);
+    XCTAssertEqual([wAdapter sectionForSectionController:sc], 0);
+}
+
 @end